Extract `sign_new_bundle()`
I was trying to abstract more common logic even further, but seemingly
much more efforts are needed. With this helper function extracted, it's
acceptable for now IMO.
